2598 fix location table + empty contact list message fix (#2602)
* fix: 2598:  Update empty location list message based on path

* fix: 2598:  Improve address cell layout and responsiveness

This commit addresses the issue where the layout of the address cell in the finder table was not responding correctly to viewport size changes. Styling has been improved to ensure elements display correctly in space constraints. Furthermore, unnecessary children props were removed from the TableCell.tsx file to clean up the component structure. The display property of .addressCellWrapper has been set to inline-flex in order to achieve better alignment and flow of the elements inside the cell. The 'showMoreLocationsIcon' has been adjusted for better visibility and interaction.