Refactor TimelineItem and EmailTimelineItem for type safety
TimelineItem now accepts a date prop instead of a user prop so that the date can be directly passed, resulting in clearer prop purposes. In the OrganisationTimeline component, added a condition to handle only InteractionEvent type events to improve runtime type safety. An InteractionEventParticipant type has been added in the getParticipant function for more accurate typing. EmailTimelineItem now takes an email prop of type InteractionEvent for safer and easier retrieval of necessary information.