fix: #2770 Fixed request url
Removed the unnecessary use of environment variable `NEXT_PUBLIC_SSR_PUBLIC_PATH` in the URL passed to `GraphQLClient`. This was not needed since the API's request path is unlikely to change and also makes the code cleaner and easier to understand. A comment was added to denote the possible need for a variable, should the request path change in the future.