Add copy-to-clipboard function and adjust UI for email composition
Introduce several updates in the EmailPreviewModal and ComposeEmail, in order to enhance the email composing and previewing process. Firstly, a copy-to-clipboard function is added to the email preview modal. Now, users can copy the current window location URL to clipboard by clicking on the copy button in the modal. The second update is about UI adjustments. The card height in the EmailPreviewModal is corrected to limit overflow, and the height of Textarea in the ComposeEmail is made more dynamic to prevent unnecessary scrolling area.

Furthermore, new FormInputs are introduced to allow users to fill in the "To" and "Subject" information of an email. Lastly, a new 'ComposeEmail.dto.ts' to handle data of new email and a git hook 'useCopyToClipboard' for clipboard functionality are both added. These changes are aimed to better suit the needs of users when handling email related activities.