Add CC and BCC options to email composer
Updated the email compose component to include CC and BCC options, implementing useOutsideClick hook to close the options when a click is made outside them. This change provides the user with more control over the recipients of an email. The 'To' and 'Subject' fields are now separated and displayed in a columnar Flex layout. EmailMetaDataEntryInput is now being used to add CC and BCC fields. The 'Reply' format in the subject line was also changed for better readability.