Fix #2823  #2824  #2814  #2815  #2816: Bug fixes  (#2828)
* fix #2816: Refactor timeline items display and improve participants' names handling

In this commit, the file EmailTimelineItem.tsx was renamed to EmailStub.tsx to better reflect its purpose. The usage across the project has been updated accordingly.
The strategy for generating participants' names in emails items was changed. This would enhance the display of names for emails when either a contact or a user is present.
Meanwhile, some unnecessary console.logs were removed from the EmailPreviewModal component. Other changes include adding a comma and margin for separating email addresses in the EmailMetaDataEntry component.
All modifications were performed to provide a better user experience and improve code clarity within the component organization structure.

* fix #2815: Improve formatting of hyperlinks in email stubs

* fix #2814: Show industry if available event when string does not match GICS options

* fix #2824 & #2823: Updated logo, removed "spaces" name from metadata