fix #2818: Archive organizations from table UI (#2893)
All references to deleting an organization have been replaced with an archive option. The mutation to archive organizations has been used in place of the old mutation for deletion. Necessary changes to make this transition have been made in the GraphQL schema and its corresponding generated types in the codebase. This prevents the total loss of information when a user opts to remove an organization. The appropriate changes were also made to checkboxes in the interface, adding new states and behaviors. This adds a more user-friendly option for removal and improves data persistence.