Improve user interface of email composer
Modified user interface for the email composer in the Timeline component. Decreased the maximum width from 100% to 90% to prevent overflowing, and adjusted the overflow properties for better scrolling behaviour.

Also enhanced the FormInput within the EmailSubjectInput. Removed the border on hover, active, focusVisible, and focus states to improve the input field appearance, and increase form clarity and overall usability. Removed button opacity setting as it was unnecessary.