Remove unused state and refactor email composition logic
The state variable 'isTextAreaEditable' was removed as it was not used anywhere in the component. Its related onFocus and onBlur functionalities were accordingly removed as well. Also, the logic for setting the mode and the default content during email composition was refactored. Now, whenever a new mode is selected, the email content field is emptied. This is to ensure that the content from a previously selected mode (e.g. REPLY_MODE) does not stay in the ComposeEmailDto object when we move to a different mode (e.g. FORWARD_MODE) thereby improving usability.