#2908 Timeline UI bugs (#2910)
* Add new Stamp icon component

* Update EmailStub component and improve readability

Replaced webp icon with a Stamp icon. The EmailStub's Card component's background color was changed to white for a more clean look. The function getEmailParticipantsName now only takes the first 'sentBy' participant to simplify view. Font weight was changed to 'semibold' to make text more readable. Reduced the size of CardFooter component.

* Update background colors and padding in UI components

The background color properties of UrlInput, TabsContainer, AboutPanel, and MainSection components have been updated to enhance UI contrast and readability. Also, padding-top in CardHeader in the MainSection has been increased for better visual spacing. These changes are part of the effort to improve user experience and align with the new design system.

* Update api key name

* Change OrganizationSidenav background color

The background color for OrganizationSidenav was changed from white to gray.25. This change was made to improve the contrast and readability for users, providing a better user experience.

* Refactor email components for better readability and cleaner ui

Code adjustments were made to the email components in the timeline events for and improved user experience. The layout for the email metadata entry, email preview modal, and compose email sections were improved for readability by ensuring overflow is hidden and text does not extend beyond its container.Lastly, adjustments to margin and padding in participant sections to avoid overlaps.

* Improve user interface of email composer

Modified user interface for the email composer in the Timeline component. Decreased the maximum width from 100% to 90% to prevent overflowing, and adjusted the overflow properties for better scrolling behaviour.

Also enhanced the FormInput within the EmailSubjectInput. Removed the border on hover, active, focusVisible, and focus states to improve the input field appearance, and increase form clarity and overall usability. Removed button opacity setting as it was unnecessary.

* Remove unused state and refactor email composition logic

The state variable 'isTextAreaEditable' was removed as it was not used anywhere in the component. Its related onFocus and onBlur functionalities were accordingly removed as well. Also, the logic for setting the mode and the default content during email composition was refactored. Now, whenever a new mode is selected, the email content field is emptied. This is to ensure that the content from a previously selected mode (e.g. REPLY_MODE) does not stay in the ComposeEmailDto object when we move to a different mode (e.g. FORWARD_MODE) thereby improving usability.