fix: People section header should be stick to top
Changes made to the "Box" and "VStack" components in the PeoplePanel ensure that the main header section containing the 'People' text and 'New Person' button stick to the top when scrolling through the list. Padding adjustments were also made for aesthetic consistency and better user experience. This was done to make permanent access to main controls possible throughout browsing the page and to make the layout visually more appealing.