Refactor delete confirmation in organization list
This commit refactors the delete confirmation dialog in the organization list component. Rather than manually creating the modal and handling the delete actions within the list component, a newly created `ConfirmDeleteDialog` component is now used. This new component not only declutters the list component ensuring cleaner and more maintainable code, but also allows for reusability of delete confirmations across the app. The list of organization IDs marked for deletion was moved to the list's state to ensure it persists in the case of unmounting the confirmation dialog. Also, a `loadingText` prop of 'Deleting' was added to the delete button in the `ConfirmDeleteDialog` component to provide user feedback during processing.