fix: #2915 People tab related bug fixes (#2922)
* fix 2915: Adjusted PeoplePanel overflow styling to make the list fully viewable. Added a height property of '100%' to the Box component in PeoplePanel.tsx to ensure the list of people occupies the full height and doesn't overflow.

* fix: The trash icon should only be red on hover. Otherwise grey.
 Clicking the check mark should collapse the card

* fix:  The Notes field border should not turn grey on hover when it is already focused

* fix: Reduced the spacing between Role and Company name to 8px

* fix: Fixed incorrect avatar icon

* Disable autocomplete for form inputs

Autocomplete feature has been disabled for all form inputs within the 'spaces' app. This was done to enhance privacy and security for our users, as sensitive user data such as passwords and other protective entries won't be stored in the browser. Additionally, hiding certain styles related to LastPass was necessary to avoid interference with our custom UI.

* fix: People section header should be stick to top

Changes made to the "Box" and "VStack" components in the PeoplePanel ensure that the main header section containing the 'People' text and 'New Person' button stick to the top when scrolling through the list. Padding adjustments were also made for aesthetic consistency and better user experience. This was done to make permanent access to main controls possible throughout browsing the page and to make the layout visually more appealing.

* Updated button styles in PeoplePanel

* Refactor ConfirmDeleteDialog component

BackgroundPattern.tsx was removed and its graphics were replaced with a background image. Changes were made to ConfirmDeleteDialog component making it more reusable; label, confirmButtonLabel and cancelButtonLabel were added as props, thus making it less specific to deleting contacts. The component was additionally moved to @ui/presentation/Modal/ for better organization. A minor format fix was done in EmailParticipantSelect.tsx to improve readability.

* Refactor delete confirmation in organization list

This commit refactors the delete confirmation dialog in the organization list component. Rather than manually creating the modal and handling the delete actions within the list component, a newly created `ConfirmDeleteDialog` component is now used. This new component not only declutters the list component ensuring cleaner and more maintainable code, but also allows for reusability of delete confirmations across the app. The list of organization IDs marked for deletion was moved to the list's state to ensure it persists in the case of unmounting the confirmation dialog. Also, a `loadingText` prop of 'Deleting' was added to the delete button in the `ConfirmDeleteDialog` component to provide user feedback during processing.

* Fix nav on item click (fix attempt)

* Change api key

The environment variable for the Communication API key was misnamed. It has been updated to be more specific to its use in mailing. Replaced `COMMS_API_KEY` with `COMMS_API_MAIL_API_KEY` to accurately reflect its purpose and improve code clarity.