fix #3000: Updated API key header key and removed unnecessary headers… (#3001)
* fix #3000: Updated API key header key and removed unnecessary headers object in post request

This commit modifies the header key from 'X-Openline-API-KEY' to 'X-Openline-Mail-Api-Key' to streamline the variable naming convention with the rest of the application. This change was done in spaces/middleware.ts. Additionally, the headers object in ComposeEmail.tsx is removed from the post request, because the necessary headers have already set in the middleware. Also added a console log for debugging purposes.

* chore: cleanup