#3029 UI improvements (#3023)
* Adjusted organization timeline, actions and events padding and margin for better view. Also, added a console log in the OrganizationTimeline.tsx for debugging the timelineEmailEvents data. These changes are to improve the user's experience in viewing the timeline and related emails.

* Add console log for request debug and correct API call.

Added console log in middleware.ts for debugging "request" and "requestHeaders" in mail API. Also made minor modifications in ComposeEmail.tsx - added trailing slash in API endpoint URL and commented out "setIsSending(true)" line for the time being. The adjustments were necessary for better request handling and to fix the issue of sending emails.

* chore: Cleanup.

* Uncomment setIsSending in ComposeEmail.tsx

Uncomment the setIsSending function call in the ComposeEmail.tsx file. This ensures the UI reflects the state that the email is being sent to the user, providing appropriate user feedback during the email sending process.