Improve layout of email composer view
Layout of the email composer view was modified to enhance UX. The main change is setting each Flex or Box that wraps an interactive element to 100% width. This was done to make the clickable area of each element more intuitive and user friendly. Details of recipients in the to, cc, and bcc fields now take up the entire width of their container, making it clearer where to click to interact with these fields. Several other slight layout adjustments were also made.