3038 UI fixes, stamp quality, mail ui improvements (#3040)
* Improve layout of email composer view

Layout of the email composer view was modified to enhance UX. The main change is setting each Flex or Box that wraps an interactive element to 100% width. This was done to make the clickable area of each element more intuitive and user friendly. Details of recipients in the to, cc, and bcc fields now take up the entire width of their container, making it clearer where to click to interact with these fields. Several other slight layout adjustments were also made.

* "Add 'To' field and remove unused code in ComposeEmail component"

The 'To' field was added to the 'ComposeEmail' component to handle more recipients while composing an email. Also, an unused check 'if (!textEmailContent) return' was removed to clean up the code. Moreover, an expected TypeScript error was removed, and the 'reply all' functionality was updated to include the 'To' field recipients.

* Update cursor style on hover in SyncSelect

The change adds a '_hover' property in the container style of the SyncSelect component to improve user experience. Now, when a user hovers over the selectable area, the cursor style will update to 'pointer', indicating an interactive element.

* Update flex attributes in Timeline components

This update removes specific comments and height constraints from EmptyTimeline.tsx and updates the height calculation in OrganizationTimeline.tsx. This was done to optimize the viewing experience across different screen sizes while maintaining the visual design. The update ensures that the timeline is always fully viewable without unnecessary scrolling and keeps the overall layout consistent.

* "Improve organization timeline view for empty states

The organization timeline view has been simplified for empty states. Instead of returning a complete rendered component upon encountering a null state, return only the EmptyTimeline component. This reduces code duplication and increases maintainability. The TimelineActions component is also imported and utilized within the EmptyTimeline file, making its rendering more consistent regardless of the state of the timeline."

* Uploaded higher quality stamp