Update email sending feedback
Updated the feedback messages that users get when they send an email. If the email is sent successfully, a toast success message will show up. If there are errors, the error reason and subject will be included in the toast error message. This would provide a better experience for users, as they will have a clearer idea about the result and possible issues.