Update toast close timing and padding for email response
This commit changes the autoClose property for the ToastContainer in layout.tsx from 3000000 to 8000 to create a more responsive user interface. It also adjusts the padding on the left (pl) and right (pr) for the email response buttons in EmailResponseModeChangeButtons.tsx to enhance the visual appeal. These changes aim to improve the user's experience while interacting with notifications and the email response interface.