fix 3053: Remove unnecessary check in ComposeEmail.tsx
This commit removes an unnecessary mode comparison in the ComposeEmail component which was causing redundancy. The removed condition check was blocking the recomposition of default email values when the email mode remained the same. It turns out this check was excessive and redundant. Removing this will improve performance and simplification of the code.