fix #3057 Replace translateY with transform in EmailResponseModeChang… (#3058)
* fix #3057 Replace translateY with transform in EmailResponseModeChangeButtons.tsx

The translateY property was replaced with the transform property for enhanced compatibility with more browsers and to adhere to CSS standards. The transform property is more widely supported and hence, would improve the application's overall accessibility.

* Adjust padding in TimelineActions component

Updated the padding at the bottom of the TimelineActions component. This change enhances visual balance within the interface, improving readability and user experience. Specifically, the padding after the component (when not shown) is increased from 4 to 8, and a redundant padding bottom within the component is removed.

* Reduced spacing between subject and email text

* Fix error white fetching new timeline data