refactor: #3064 Button theme update  (#3065)
* fix 3064: Update button theming

In this change, the inline styles for the button in PeoplePanel.tsx were removed. These stylings were instead included in the theme configuration for the button component - Button.theme.ts. This organizes our code better and allows for reusability of similar button styles across the application.

* chore: Cleanup
Refactor and simplify button CSS styling in PeoplePanel and OrganizationTimeline

The CSS styling for the buttons in the PeoplePanel and OrganizationTimeline components has been heavily rewritten. Most of the removed code segments controlled the colours of certain button states (_hover, _focus, _focusVisible). This redundant customization was removed to streamline the code, improve readability and maintenance.