fix #3069, #3083, #3084 Email validation, copy change, separated about and ppl tab query (#3086)
* (chore) Update theme for gray color scheme buttons

* fix #3083: Copy change

* refactor #3084: Remove contacts fields from organization query
Removed 'contacts' fields from the 'organization' query in the files 'organization.generated.ts' and 'organization.graphql'. This helps reduce the load on the server and improve application performance by making the request more lightweight and fetching only the most essential organization details. The contacts data is now fetched more efficiently through the new 'organizationPeoplePanel' query.

Add OrganizationPeoplePanel query and simplify PeoplePanel
Created organizationPeoplePanel query to efficiently fetch the essential contact details from the server. As a result, the PeoplePanel and ContactCard components were simplified by eliminating various mutations that would have been called in previous implementation. Additionally, some functionalities and states were moved to the ContactCard component for better encapsulation and code readability. This better aligns with the separation of concerns principle. The changes reduce load on server and improve application performance.

* chore: Update SimpleValidationIndicator functionality and styling

This commit involves the update of the SimpleValidationIndicator component. The main changes include the addition of a new overlay popover feature for error messages and changing the styling of the validation indicator. Import paths were also restructured and InlineLoader was moved to a new location. The presence of the validation signal no longer depends on the edit mode. The changes were made to improve user experience by providing a clearer visualization of validation states and errors.

* fix: 3069: Add email validation to contact card
Improved close contact card
This change introduces the capability to validate emails in real-time via GraphQL client in the contact card component. A new 'useTenantName' hook was added for fetching tenant names, which comes as a new requirement for email validation. Code refactoring is also done to adjust the organization of related components. The file 'ContactCard' has been moved into its own folder with the same name, and a new 'EmailValidationMessage' file has been created to hold the validation logic, making the code structure more maintainable and straightforward. This entire change is initiated to make our application more user-friendly by providing immediate feedback on the input email format.

* chore: Remove unnecessary console log in useTenantName hook