fix: 3168 UI feedback for account panel (#3172)
* fix 3168: Update styling for date-picker

The diff shows updates to the date-picker's styling. The changes include removal of the styles applied to the current date's tile, and the addition of a rule to remove text decorations from elements within the `.react-date-picker` class. Also, adjustments are made to the background color and border color of tiles marked as having active states.

* fix 3168: Changed the label text from 'Renewal likelihood' to 'Renewal forecast' providing clearer communication for users.

* Remove condition for displaying 'Reason for change' field in RenewalForecastModal.

* Added organization name to renewal forecast

* Updated margin top value in RenewalForecastModal

Decreased margin top value of the label for 'Reason for change' in the renewal forecast modal.

* Added currency formatting utility methods and update modal components & automatically focus amount input in RenewalForecastModal

* Updated BillingDetailsCard placeholder text

* Adjusted gap between elements in BillingDetailsCard

The spacing between the FormCurrencyInput and FormSelect elements in the BillingDetailsCard component in 'spaces' app was adjusted to improve layout and readability. This change was prompted by a design review and user feedback indicating that the previous layout led to visual clutter and was difficult to navigate. The use of the 'gap' attribute ensures consistent spacing between card items.

* Copy change.
The label text for the FormCurrencyInput within the BillingDetailsCard component was updated from 'Billing amounts' to 'Billing amount'.

* Update placeholder and remove currency formatting in RenewalForecast

The placeholder in the RenewalForecastModal was updated from a hard-coded value to a more generic placeholder. The currency formatting functions were also removed to simplify input handling. Users can now input amounts without specific currency format restrictions.

* Update RenewalForecast component for display improvements
Renewal Forecast heading is updated to have a consistent letter casing across the app.
The amount format now has a minimum fraction digit of 0, to hide decimals.
The reason text is also limited to 2 lines after which it gets truncated

* Fixed resetting the amount to null and display value

* Added spacing between title and amount for renewal forecast

* Fix missing return statement in default case

In the 'RenewalLikelihood.tsx' file, a missing return statement in the default case for switch function of likelihood, was causing undesired output. This has been fixed to return 'Not set'.

* Updated account tab and renewal likelihood icon

* Added organization name to RenewalLikelihood components

Added the organization name to the RenewalLikelihood, RenewalLikelihoodModal and AccountPanel components. This change was necessary to personalise the renewal likelihood warnings and information, instead of referring to the generic 'Acme Corp'. The organization name is now dynamically displayed in these sections."

* Changed zero label color to gray.700

* chore: Copy change

* Unified placeholder color to gray.400