fix 3230 Fixed reply all in email, participant display, email spacing and overflow in modal
This commit includes multiple improvements in the email component. Firstly, participant tags were changed to React Fragments to ensure that labels and separating commas render correctly. The BCC button that was previously commented out has been restored to improve email composition capabilities. The 'to' field in the reply-all functionality has also been adjusted to correctly retrieve participant names and emails. The maximum height of the email compose window was adjusted for better visibility in modal view. Lastly, an unused import in the RenewalLikelihood.tsx file was removed to clean up the code.