refactor: Replace "activeTab" with "lastPositionParams" for better clarity
ActiveTab name was implying a specific use case for navigation tabs but then it was also used for storing last position in the app components. Therefore, it's now refactored to "lastPositionParams" which better describes its purpose across the application. This change is made to enhance code readability and maintainability in the long term.