feat 3391 Remember previewed event (organization event) (#3392)
* feat #3391: Add last active tab position to local storage

This commit introduces a feature that retains the last active tab position in an organization's timeline. The 'useLocalStorage' method was imported and used to store the last active position. This update will improve user experience by reducing the need for re-navigation.

* refactor: Replace "activeTab" with "lastPositionParams" for better clarity

ActiveTab name was implying a specific use case for navigation tabs but then it was also used for storing last position in the app components. Therefore, it's now refactored to "lastPositionParams" which better describes its purpose across the application. This change is made to enhance code readability and maintainability in the long term.