fix #3463: Update last active position on modal close (#3464)
The TimelineEventPreviewContext closing function is updated to save the last active position. This change keeps track of the page's last active state before closing the modal, ensuring that the user is brought back to the correct point after the modal closure and navigating back / forth to given org details page.