fix #3461: Fix validation for closing email preview modal (#3462)
Added additional check functions 'checkPristine' and 'checkEmpty' to enhance the validation when closing the email preview modal. This was implemented to ensure that the confirmation dialog only shows when it should. For instance, when participant fields are exploited and are not empty, or when the subject is not pristine with a length, and the form is not empty. The onClose function is only invoked when the form is pristine or the confirmation dialog doesn't need to show.