(fix|COS-232): Allow users to close the dialog by pressing the 'Esc' key
This change was made to improve the user experience by providing a quick escape route in case the user changes their mind. Now, when the user is faced with the delete confirmation dialog, they can simply press the 'Esc' key to close this dialog. This functionality adheres more closely to the default user expectations of a dialog behavior.