(feat|COS-237): Added externalLink support for LogEntry, Hubspot and Salesforce logos, script for logo generation, (#3804)
Added Hubspot and Salesforce logos to ui/media/logos. A script for automatic logo generation from svg files was also added. ExternalLinks property was added to the LogEntry objects, which helps connect the data to more sources. Added a LogEntryExternalLink component, that takes an external system link and generate an appropriate icon and a link to the source.

This allows external sources like Hubspot and Salesforce to be integrated into the system providing users access to other platforms increasing overall usability.