(fix|COS-240) UI improvements for log entry feature (#3806)
- Refined HTML content rendering for consistent styling
- An update has been made in the HTMLContentRenderer.tsx in filtering certain attributes namely 'font-size', 'font-family', and 'font-color' from inline styles. This change was needed to ensure uniformity in styling across all elements, as these specific font styling attributes were causing inconsistencies in font presentation. Font attributes should be inherited from parent element
- Updated placeholder style in remirror-editor
- Adjusted styling for new tag options dropdown in Timeline
This change gives a new pre-selection UI for tag options dropdown in Timeline. When a new tag (which is not part of the suggested list) is being added, the dropdown becomes hidden. This prevents irrelevant options from being displayed, thus providing a cleaner user interface experience.
- Added 'emails' field to the GraphQL query in 'getTimeline.graphql' and updated respective generated typescript file. In the UI layer, updated 'LogEntryPreviewModal.tsx' to display author's email in the tooltip.
This change was made to provide more information about the author in the log entry preview dialog, enhancing the user experience and improving the detailed examination capability in the application.