(fix|COS-49): Refactor timeline cache update and email sending process
The changes in these files mainly aim to refactor and optimize the process of updating the timeline cache and the handling of the email sending process. A new custom hook 'useUpdateCacheWithNewEvent' was added to handle the updating of the timeline caching process which was formerly implemented within specific components, hence improving code modularity. A 'generateTimelineEvent' function was also created to create a separate timeline event, making the handleSendEmail function more manageable