(fix|COS-407): Improve HTML content rendering by cleaning and formatting - fix bug in email preview ui (#3953)
* (fix|COS-407): Improve HTML content rendering by cleaning and formatting - fix bug in email preview ui

The changes are introduced to the "HtmlContentRenderer.tsx" to improve the rendering of HTML content. Broadly, there are two key changes associated with this commit:

1. Clean HTML function: A new function "cleanHTML" is created to efficiently clean the HTML content. This includes removing unnecessary tags like "!doctype", "html", "title", "body", "meta" and the content within style tags for optimal rendering.

2. The handling of HTML attributes is improved to exclude ‘style’ attributes while preserving others — aiding in better content presentation.

Overall, these changes provide better UI/UX for log entries while ensuring smoother HTML content rendering.

* Refactor HTML sanitization logic in HtmlContentRenderer

This commit replaces the custom HTML sanitization function with the sanitize-html library in HtmlContentRenderer.tsx. The custom function was removing certain tags from the HTML content but it was not sufficient for all cases and caused UI issues. The sanitize-html library provides a more robust and secure implementation. It is configured to allow specific tags and attributes that we accept in our HTML content, while sanitizing the rest.

* Updated styling properties in HtmlContentRenderer

This commit updates the padding, display, border, and borderRadius properties in the HtmlContentRenderer component. The padding has been increased to improve the readability and aesthetics of the content. The display was set to 'block' to fix a layout issue where content was not properly aligned. The border and borderRadius were added to make the content more distinct and visually appealing. A redundant div styling was removed to simplify the styling rules.